+++ src/game/Melee_Greatclub.java
@@ -10,7 +10,7 @@ package game;
  * @author nerdi
  */
 public class Melee_Greatclub extends Melee {
-    private final static long itemID = 11003;
+    private final static int itemID = 11003;
     private final static String name = "Gratclub";
     
     private final static boolean twoHanded = true;
