+++ src/game/Melee_Handaxe.java
@@ -10,7 +10,7 @@ package game;
  * @author nerdi
  */
 public class Melee_Handaxe extends Melee {
-    private final static long itemID = 11004;
+    private final static int itemID = 11004;
     private final static String name = "Handaxe";
     
     private final static boolean twoHanded = false;
