+++ src/game/Melee_Mace.java
@@ -10,7 +10,7 @@ package game;
  * @author nerdi
  */
 public class Melee_Mace extends Melee {
-    private final static long itemID = 11006;
+    private final static int itemID = 11006;
     private final static String name = "Mace";
     
     private final static boolean twoHanded = true;
